requirements.txt
mecab-python3
mecab-ko-dic
eunjeon
requests
ko-pron
jaconv